Change Chapter 04

+++ BT_04_DAMTUANPHAT_N21DCCN161_D21CQCN02-N/Bai7.cpp
@@ -11,10 +11,10 @@ int factorial(int n) {
     return n*factorial(n-1);
 }
 
-int  tongNDGT(int n) {
-    int sum = 0;
+float  tongNDGT(int n) {
+    float sum = 0;
     for(int i = 1; i <= n; i++) {
-        sum += 1/factorial(i);
+        sum += 1/ (float) factorial(i);
     }
 
     return sum;
